requirements.txt
pystemmer==2.0.1
